+++ packages/plugin-oauth-google/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @swr-login/plugin-oauth-google
 
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @swr-login/core@0.8.0
+
 ## 0.2.5
 
 ### Patch Changes